adds: clean up of server codebase into dedicated files for easier development
